Number movement tiles and use speed
This change introduces a movement controller for the player characters.
The movement controller gives the CursorController a speed that
maximizes the amount of green tiles placed. Each green tile is also
numbered according to which order in the movement it takes place.

The cursor controller was also changed to organize these movements with
a stack. The player controller will inherit a full movement phase and
pop each movement accordingly.