Add turn selector and sprites for move phase
This changes adds more components to the player's turn action phase. The
action phase consists of creationg a movement, action, and turn plan.
The user should be able to plan out each of those actions for one unit
without influencing the rest of the game state. After a character's
action is locked they should be hoverable to view all of their selected
movements. This patch adds the sprites and selector functionality.

The second part of this patch will focus more on refactoring to allow
the hovering option to be easily accessible. Currently it's a challenge
as I've written too much spaghetti code as I worry about the game jam
deadline.