Smoothed out player action selection
Made the player action selection sequence less buggy and more linear.
The player is able to select the movement, action, and turn of each of
their individual characters. During the aciton and turn sequences the
player is able to cancel the sequence overall. Hovering over a unit with
a decided movement action shows all of their action toggles.