Add input protection and movement turning
Adds safeguards to ensure the player can't make incorrect movements.
When planning an action they shouldn't be able to interact with anything
aside from the buttons. This should also stop the player from starting
the movement phase while in the middle of something. It also provides a
bit more cleanup to ensure every character gets reset inbetween rounds.

This patch also makes it so characters take their turn action at the end
of a movement.