Added camera scrolling
Added the ability to pan the mouse near the edge of the screen for
camera scrolling. The camera will stop scrolling near the edge to ensure
players don't move the camera outside of the map.

There's also a dumb thing that centers the camera on a selected unit,
but that'll be removed when I'm more awake because it's too confusing
and not useful.