Differentiate enemy units and pause on collision
Changed some settings to differentiate how enemies can be interacted
with. The player is unable to move them, and a new AI controller is
responsible for managing their AI. Also updated the movement and flow
controller scripts to force all units to sleep when two units collide.
This logic will later be updated to introduce the combat system.