Added action handling to combat phase
Added a handler in the CharacterStats script that allows for the
handling of action buttons. With this patch characters are able to rush
or fortify. Class-specific actions will also live in this script
(probably all in the same script due to lack of time).

This change also updates some UI elements. Sorry for having bad commit
messages.